Feat: 3차 QA 작업 해결 1.
