+++ .gitignore
@@ -23,6 +23,7 @@
 
 # production
 /build
+/public
 
 # misc
 .DS_Store
