+++ src/app/events/[id]/page.tsx
@@ -5,6 +5,8 @@ import { use, useState } from 'react';
 import Image from 'next/image';
 import Link from 'next/link';
 
+import { handleShare } from '@/app/utils/clipboard';
+
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
 
@@ -36,69 +38,6 @@ export default function EventDetailPage(props: { params: Promise<{ id: string }>
     // 실제로는 API 호출하여 북마크 상태 변경
     toast(isBookmarked ? '북마크가 해제되었습니다' : '북마크에 추가되었습니다');
   };
-
-  const fallbackCopyToClipboard = (text: string) => {
-    const textArea = document.createElement('textarea');
-    textArea.value = text;
-    textArea.style.position = 'fixed';
-    textArea.style.top = '-9999px';
-    document.body.appendChild(textArea);
-
-    textArea.focus();
-    textArea.select();
-
-    try {
-      const successful = document.execCommand('copy');
-      if (successful) {
-        toast('링크가 복사되었습니다', {
-          description: 'fallback 복사 테스트.!!!!!링크 공유됨',
-        });
-      } else {
-        throw new Error('execCommand 실패');
-      }
-    } catch (err) {
-      console.error('Fallback 복사 실패:', err);
-      toast.error('복사 실패!!!!!!!', {
-        description: '브라우저 설정을 확인해주세요.',
-      });
-    }
-    document.body.removeChild(textArea);
-  };
-
-  const handleShare = async () => {
-    const url = window.location.href;
-
-    if (!navigator.clipboard) {
-      fallbackCopyToClipboard(url);
-      return;
-    }
-
-    try {
-      await navigator.clipboard.writeText(url);
-      toast('링크가 복사되었습니다', {
-        description: '친구들에게 공유해보세요!',
-      });
-    } catch (err) {
-      console.error('복사 실패!!!!!!!!222Clipboard API 문제:', err);
-      fallbackCopyToClipboard(url);
-    }
-  };
-
-  // const handleShare = () => {
-  //   // 현재 URL을 클립보드에 복사
-  //   const url = window.location.href;
-  //   navigator.clipboard
-  //     .writeText(url)
-  //     .then(() => {
-  //       toast('링크가 복사되었습니다', {
-  //         description: '친구들에게 공유해보세요!',
-  //       });
-  //     })
-  //     .catch((err) => {
-  //       console.error('클립보드 복사 실패:', err);
-  //     });
-  // };
-
 
   // 날짜 포맷팅 함수
   const formatDate = (dateString: string) => {
